Extract Myspace shared item models
